pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "decky-plugin-store"
version = "0.1.0"
description = "Plugin Store backend for Decky"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.9.1"
alembic = "^1.8.1"
aiosqlite = "^0.17.0" # for async sqlite in sqlalchemy
asgiref = "^3.5.2"
discord-webhook = {version = "^0.17.0", extras = ["async"]}
fastapi = "^0.85.1"
python-multipart = "^0.0.5"  # for multipart forms in fastapi
sqlalchemy = "^1.4.41"
uvicorn = "^0.20.0"
limits = {extras = ["redis"], version = "^3.6.0"}
redis = "^5.0.1"
asyncpg = "^0.29.0" # for async postgres in sqalchemy

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.2"
httpx = "^0.23.0"
isort = "^5.10.1"
mypy = "^0.991"
pytest = "^7.2.0"
pytest-asyncio = "0.19.0"
pytest-env = "^0.6.2"
pytest-freezegun = "^0.4.2"
pytest-lazy-fixture = "^0.6.3"
pytest-mock = "^3.8.2"
sqlalchemy2-stubs = "*"
types-pytest-lazy-fixture = "^0.6.3.2"

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
(
    /(
        \.eggs
      | \.git
      | \.hg
      | \.mypy_cache
      | \.tox
      | \.venv
      | _build
      | buck-out
      | build
      | dist
    )/
)
'''

[tool.flake8]
max-line-length = 120

[tool.isort]
line_length = 120
order_by_type = false
profile = "black"
src_paths = ["plugin_store"]

[tool.mypy]
plugins = [
#    "sqlmypy",
    "sqlalchemy.ext.mypy.plugin",
]

[[tool.mypy.overrides]]
module = ["discord_webhook", "limits"]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = "/app/plugin_store"
env = [
    "DB_URL=sqlite+aiosqlite:///:memory:",
    "SUBMIT_AUTH_KEY=deadbeef",
]
